AUTH - [FIX] auth details : move to core
